Stage Circular Dependency correction, Class Stubs needed for Stage added, Documentation and Minor corrections
In an effort to duplicate the same structure as flash api, a circular dependency needed to be corrected so Stage class can still extend DisplayObjectContainer.  Sprite now creates the Stage instance and passes it to DisplayObject as a static variable that is available when instantiating DisplayObject on it's own or extended.  Additionally some classes include documentation to further the duplication of the Flash API.  Minor adjustments have also been made to for SWC generation.